new commit

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -4,6 +4,8 @@ import com.ust.pos.dto.UserDto;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -20,6 +22,7 @@ public class UserController {
 
     @GetMapping("/list")
     public String home(Model model) {
+
         model.addAttribute("users", userService.findAll());
         return "user/list";
     }
@@ -28,7 +31,7 @@ public class UserController {
     public String update(Model model, @RequestParam String username, @ModelAttribute UserDto userDto) {
         UserDto response = userService.findByUserName(username);
         model.addAttribute("userDto", response);
-        model.addAttribute("roles",roleService.findAll());
+        model.addAttribute("roles", roleService.findAll());
         return "user/user";
     }
 
@@ -44,7 +47,19 @@ public class UserController {
 
     @GetMapping("/delete")
     public String delete(Model model, @RequestParam String username) {
-        userService.delete(username);
-        return "redirect:/user/list";
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication != null) {
+            String loggedInUser = authentication.getName();
+
+            userService.delete(username);
+
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return "redirect:/login";
+            }
+        }
+            return "redirect:/user/list";
     }
+
 }
